refactor: project file system and lint private rules move
