+++ .eslintrc.js
@@ -11,7 +11,6 @@ module.exports = {
         "sourceType": "module"
     },
     "rules": {
-        "no-console": "off",
         "import/extensions": "off"
     }
 }
